requirements.txt
torch>=2.6.0
fbgemm_gpu>=1.1.0
torchrec>=1.1.0
gin_config>=0.5.0
pandas>=2.2.0
tensorboard>=2.19.0
pybind11
